karim_v4

+++ routes/api/orientationinvitations.js
@@ -4,6 +4,8 @@ const mongoose = require('mongoose')
 
 const OrientationInvitation = require('../../models/OrientationInvitation')
 const validator = require('../../validations/orientationinvitationValidations')
+const ObjectId=require('mongodb').ObjectID;
+mongoose.set('useFindAndModify',false);
 
 router.get('/', async (req,res) => {
     const orientationinvitations = await OrientationInvitation.find()
@@ -33,14 +35,17 @@ router.post('/', async (req,res) => {
 // Update an OrientationInvitation
 router.put('/:id', async (req,res) => {
     try {
-     const id = req.params.id
-     const orientationinvitation = await OrientationInvitation.findOne({id})
-     if(!orientationinvitation) return res.status(404).send({error: 'Orientation Invitation does not exist'})
-     const isValidated = validator.updateValidationOrientationInvitation(req.body)
-     if (isValidated.error) return res.status(400).send({ error: isValidated.error.details[0].message })
-     const updatedOrientationInvitation = await OrientationInvitation.updateOne(req.body)
-     res.json({msg: 'Orienation Invitation updated successfully'})
-    }
+        if(ObjectId.isValid(req.params.id)){
+            const updatedOrientationInvitation = await OrientationInvitation.findByIdAndUpdate({_id: req.params.id}, req.body)
+            if (!updatedOrientationInvitation) return res.status(404).send({error: 'Orientation Invitation does not exists' })
+            res.json({msg: 'Orientation Invitation updated successfully'})
+           }
+           else {
+               return res.status(404).send({error: "not an Orientation Invitation id"})
+           }
+        }
+    
+    
     catch(error) {
         console.log(error)
         return res.status(404).send({error: 'Orientation Invitation does not exist'})
@@ -49,9 +54,15 @@ router.put('/:id', async (req,res) => {
 
  router.delete('/:id', async (req,res) => {
     try {
-     const id = req.params.id
-     const deletedOrientationInvitation = await OrientationInvitation.findByIdAndRemove(id)
-     res.json({msg:'Orientation Invitation was deleted successfully', data: deletedOrientationInvitation})
+        if(ObjectId.isValid(req.params.id)){
+            const id = req.params.id 
+            const deletedOrientationInvitation = await OrientationInvitation.findByIdAndRemove(id)
+            if (!deletedOrientationInvitation) return res.status(400).send({ error: 'Orientation Invitation does not exists' })
+            res.json({msg: 'Orientation Invitation was deleted successfully', data: deletedPartner})
+           }
+           else {
+               return res.status(404).send({error: "not an Orientation Invitation id"})
+           }
     }
     catch(error) {
         console.log(error)
